+++ driver-core/src/main/java/com/datastax/driver/core/TableMetadata.java
@@ -18,6 +18,9 @@ package com.datastax.driver.core;
 import java.util.*;
 import java.util.regex.Pattern;
 
+import com.google.common.base.Predicate;
+import com.google.common.collect.Iterables;
+
 import org.codehaus.jackson.JsonFactory;
 import org.codehaus.jackson.map.ObjectMapper;
 
@@ -47,22 +50,41 @@ public class TableMetadata {
     private final KeyspaceMetadata keyspace;
     private final String name;
     private final List<ColumnMetadata> partitionKey;
-    private final List<ColumnMetadata> clusteringKey;
+    private final List<ColumnMetadata> clusteringColumns;
     private final Map<String, ColumnMetadata> columns;
     private final Options options;
+    private final List<Order> clusteringOrder;
+
+    /**
+     * Clustering orders.
+     * <p>
+     * This is used by {@link #getClusteringOrder} to indicate the clustering
+     * order of a table.
+     */
+    public static enum Order {
+        ASC, DESC;
+
+        static final Predicate<Order> isAscending = new Predicate<Order>() {
+            public boolean apply(Order o) {
+                return o == ASC;
+            }
+        };
+    }
 
     private TableMetadata(KeyspaceMetadata keyspace,
                           String name,
                           List<ColumnMetadata> partitionKey,
-                          List<ColumnMetadata> clusteringKey,
+                          List<ColumnMetadata> clusteringColumns,
                           LinkedHashMap<String, ColumnMetadata> columns,
-                          Options options) {
+                          Options options,
+                          List<Order> clusteringOrder) {
         this.keyspace = keyspace;
         this.name = name;
         this.partitionKey = partitionKey;
-        this.clusteringKey = clusteringKey;
+        this.clusteringColumns = clusteringColumns;
         this.columns = columns;
         this.options = options;
+        this.clusteringOrder = clusteringOrder;
     }
 
     static TableMetadata build(KeyspaceMetadata ksm, Row row, Map<String, ColumnMetadata.Raw> rawCols) {
@@ -77,11 +99,12 @@ public class TableMetadata {
         boolean isCompact = isDense || !comparator.isComposite;
 
         List<ColumnMetadata> partitionKey = nullInitializedList(keyValidator.types.size());
-        List<ColumnMetadata> clusteringKey = nullInitializedList(clusteringSize);
+        List<ColumnMetadata> clusteringColumns = nullInitializedList(clusteringSize);
+        List<Order> clusteringOrder = nullInitializedList(clusteringSize);
         // We use a linked hashmap because we will keep this in the order of a 'SELECT * FROM ...'.
         LinkedHashMap<String, ColumnMetadata> columns = new LinkedHashMap<String, ColumnMetadata>();
 
-        TableMetadata tm = new TableMetadata(ksm, name, partitionKey, clusteringKey, columns, new Options(row, isCompact));
+        TableMetadata tm = new TableMetadata(ksm, name, partitionKey, clusteringColumns, columns, new Options(row, isCompact), clusteringOrder);
 
         // We use this temporary set just so non PK columns are added in lexicographical order, which is the one of a
         // 'SELECT * FROM ...'
@@ -94,14 +117,15 @@ public class TableMetadata {
                     partitionKey.set(rawCol.componentIndex, col);
                     break;
                 case CLUSTERING_KEY:
-                    clusteringKey.set(rawCol.componentIndex, col);
+                    clusteringColumns.set(rawCol.componentIndex, col);
+                    clusteringOrder.set(rawCol.componentIndex, rawCol.isReversed ? Order.DESC : Order.ASC);
                     break;
             }
         }
 
         for (ColumnMetadata c : partitionKey)
             columns.put(c.getName(), c);
-        for (ColumnMetadata c : clusteringKey)
+        for (ColumnMetadata c : clusteringColumns)
             columns.put(c.getName(), c);
         for (ColumnMetadata c : otherColumns)
             columns.put(c.getName(), c);
@@ -160,7 +184,7 @@ public class TableMetadata {
      * The order of the columns in the list is consistent with
      * the order of the columns returned by a {@code SELECT * FROM thisTable}:
      * the first column is the partition key, next are the clustering
-     * keys in their defined order, and then the rest of the
+     * columns in their defined order, and then the rest of the
      * columns follow in alphabetic order.
      *
      * @return a list containing the metadata for the columns of this table.
@@ -179,9 +203,9 @@ public class TableMetadata {
      * @return the list of columns composing the primary key for this table.
      */
     public List<ColumnMetadata> getPrimaryKey() {
-        List<ColumnMetadata> pk = new ArrayList<ColumnMetadata>(partitionKey.size() + clusteringKey.size());
+        List<ColumnMetadata> pk = new ArrayList<ColumnMetadata>(partitionKey.size() + clusteringColumns.size());
         pk.addAll(partitionKey);
-        pk.addAll(clusteringKey);
+        pk.addAll(clusteringColumns);
         return pk;
     }
 
@@ -198,13 +222,29 @@ public class TableMetadata {
     }
 
     /**
-     * Returns the list of columns composing the clustering key for this table.
+     * Returns the list of clustering columns for this table.
      *
-     * @return the list of columns composing the clustering key for this table.
-     * If the clustering key is empty, an empty list is returned.
+     * @return the list of clustering columns for this table.
+     * If there is no clustering columns, an empty list is returned.
      */
-    public List<ColumnMetadata> getClusteringKey() {
-        return Collections.unmodifiableList(clusteringKey);
+    public List<ColumnMetadata> getClusteringColumns() {
+        return Collections.unmodifiableList(clusteringColumns);
+    }
+
+    /**
+     * Returns the clustering order for this table.
+     * <p>
+     * The returned contains the clustering order of each clustering column. The
+     * {@code i}th element of the result correspond to the order (ascending or
+     * descending) of the {@code i}th clustering column (see
+     * {@link #getClusteringColumns}). Note that a table defined without any
+     * particular clustering order is equivalent to one for which all the
+     * clustering key are in ascending order.
+     *
+     * @return a list with the clustering order for each clustering column.
+     */
+    public List<Order> getClusteringOrder() {
+        return clusteringOrder;
     }
 
     /**
@@ -306,19 +346,20 @@ public class TableMetadata {
             }
             sb.append(")");
         }
-        for (ColumnMetadata cm : clusteringKey)
+        for (ColumnMetadata cm : clusteringColumns)
             sb.append(", ").append(escapeId(cm.getName()));
         sb.append(")");
         newLine(sb, formatted);
         // end PK
 
         // Options
-        if (options.isCompactStorage) {
-            sb.append(") WITH COMPACT STORAGE");
-            and(sb, formatted).append("read_repair_chance = ").append(options.readRepair);
-        } else {
-            sb.append(") WITH read_repair_chance = ").append(options.readRepair);
-        }
+        sb.append(") WITH ");
+
+        if (options.isCompactStorage)
+            and(sb.append("COMPACT STORAGE"), formatted);
+        if (!Iterables.all(clusteringOrder, Order.isAscending))
+            and(appendClusteringOrder(sb), formatted);
+        sb.append("read_repair_chance = ").append(options.readRepair);
         and(sb, formatted).append("dclocal_read_repair_chance = ").append(options.localReadRepair);
         and(sb, formatted).append("replicate_on_write = ").append(options.replicateOnWrite);
         and(sb, formatted).append("gc_grace_seconds = ").append(options.gcGrace);
@@ -330,6 +371,15 @@ public class TableMetadata {
         and(sb, formatted).append("compression = ").append(formatOptionMap(options.compression));
         sb.append(";");
         return sb.toString();
+    }
+
+    private StringBuilder appendClusteringOrder(StringBuilder sb) {
+        sb.append("CLUSTERING ORDER BY (");
+        for (int i = 0; i < clusteringColumns.size(); i++) {
+            if (i > 0) sb.append(", ");
+            sb.append(clusteringColumns.get(i).getName()).append(" ").append(clusteringOrder.get(i));
+        }
+        return sb.append(")");
     }
 
     private static String formatOptionMap(Map<String, String> m) {
@@ -370,6 +420,7 @@ public class TableMetadata {
             sb.append('\n');
         return sb;
     }
+
 
     public static class Options {