+++ driver-examples/stress/src/main/java/com/datastax/driver/stress/QueryGenerator.java
@@ -30,10 +30,16 @@ public abstract class QueryGenerator implements Iterator<QueryGenerator.Request>
         this.iterations = iterations;
     }
 
+    public abstract int currentIteration();
+
+    public int totalIterations() {
+        return iterations;
+    }
+
     public interface Builder {
         public String name();
         public OptionParser addOptions(OptionParser parser);
-        public void createSchema(OptionSet options, Session session);
+        public void prepare(OptionSet options, Session session);
         public QueryGenerator create(int id, int iterations, OptionSet options, Session session);
     }